fix: removed forward declaration in order to allow the template function to work
