+++ Source/Application.cpp
@@ -17,8 +17,6 @@
 #include "ModulePlayer.h"
 #include "ScriptFactory.h"
 
-#include "Enums/ModuleType.h"
-
 #include <ranges>
 
 constexpr int FRAMES_BUFFER = 50;
