+++ Source/Engine.vcxproj.filters
@@ -698,9 +698,6 @@
     <ClInclude Include="Enums\ModuleType.h">
       <Filter>Enums</Filter>
     </ClInclude>
-    <ClInclude Include="Enums\ModuleType_fwd.h">
-      <Filter>Enums</Filter>
-    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <Filter Include="Auxiliar">
