--- Source/Enums/ModuleType_fwd.h
@@ -1,23 +1,0 @@
-#pragma once
-
-class ModuleRender;
-class ModuleWindow;
-class ModuleInput;
-class ModuleProgram;
-class ModuleFileSystem;
-class ModuleResources;
-class ModuleScene;
-class ModuleCamera;
-class ModuleDebugDraw;
-class ModuleUI;
-class ModuleAudio;
-class ModulePlayer;
-class ModuleEditor;
-
-enum class ModuleType;
-
-template<typename T>
-struct ModuleToEnum
-{
-	const static ModuleType value;
-};
